refactor: 장바구니 아이템 카드 영역 컴포넌트로 분리
