+++ src/refactoring/components/cart/index.ts
@@ -1,0 +1,1 @@
+export { CartItemCard } from './CartItemCard';
