+++ src/refactoring/pages/CartPage.tsx
@@ -1,5 +1,6 @@
+import { CartItemCard } from '../components/cart';
 import { useCart } from '../hooks';
-import type { CartItem, Coupon, Product, Discount } from '../../types.ts';
+import type { CartItem, Coupon, Product } from '../../types.ts';
 
 interface Props {
   products: Product[];
@@ -16,10 +17,6 @@ export const CartPage = ({ products, coupons }: Props) => {
     calculateTotal,
     selectedCoupon,
   } = useCart();
-
-  const getMaxDiscount = (discounts: Discount[]) => {
-    return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
-  };
 
   const getRemainingStock = (cart: CartItem[], product: Product) => {
     const cartItem = cart.find((item) => item.product.id === product.id);
@@ -51,55 +48,12 @@ export const CartPage = ({ products, coupons }: Props) => {
             {products.map((product) => {
               const remainingStock = getRemainingStock(cart, product);
               return (
-                <div
-                  key={product.id}
-                  data-testid={`product-${product.id}`}
-                  className='bg-white p-3 rounded shadow'
-                >
-                  <div className='flex justify-between items-center mb-2'>
-                    <span className='font-semibold'>{product.name}</span>
-                    <span className='text-gray-600'>
-                      {product.price.toLocaleString()}원
-                    </span>
-                  </div>
-                  <div className='text-sm text-gray-500 mb-2'>
-                    <span
-                      className={`font-medium ${
-                        remainingStock > 0 ? 'text-green-600' : 'text-red-600'
-                      }`}
-                    >
-                      재고: {remainingStock}개
-                    </span>
-                    {product.discounts.length > 0 && (
-                      <span className='ml-2 font-medium text-blue-600'>
-                        최대{' '}
-                        {(getMaxDiscount(product.discounts) * 100).toFixed(0)}%
-                        할인
-                      </span>
-                    )}
-                  </div>
-                  {product.discounts.length > 0 && (
-                    <ul className='list-disc list-inside text-sm text-gray-500 mb-2'>
-                      {product.discounts.map((discount, index) => (
-                        <li key={index}>
-                          {discount.quantity}개 이상:{' '}
-                          {(discount.rate * 100).toFixed(0)}% 할인
-                        </li>
-                      ))}
-                    </ul>
-                  )}
-                  <button
-                    onClick={() => addToCart(product)}
-                    className={`w-full px-3 py-1 rounded ${
-                      remainingStock > 0
-                        ? 'bg-blue-500 text-white hover:bg-blue-600'
-                        : 'bg-gray-300 text-gray-500 cursor-not-allowed'
-                    }`}
-                    disabled={remainingStock <= 0}
-                  >
-                    {remainingStock > 0 ? '장바구니에 추가' : '품절'}
-                  </button>
-                </div>
+                <CartItemCard
+                  key={`cart-item-card-${product.id}`}
+                  remainingStock={remainingStock}
+                  product={product}
+                  addToCart={addToCart}
+                />
               );
             })}
           </div>
@@ -159,6 +113,7 @@ export const CartPage = ({ products, coupons }: Props) => {
           <div className='mt-6 bg-white p-4 rounded shadow'>
             <h2 className='text-2xl font-semibold mb-2'>쿠폰 적용</h2>
             <select
+              // applyCoupon, coupons, selectedCoupon
               onChange={(e) => applyCoupon(coupons[parseInt(e.target.value)])}
               className='w-full p-2 border rounded mb-2'
             >
